Update lecture_1: ExpMod_naive <change data type>

+++ practice/lecture/lecture_1/ExpMod_naive.cpp
@@ -4,7 +4,7 @@ int main(){
     unsigned long long a, b, c=1e9 +7, ans=1;
     cin>>a>>b;
     unsigned long long a_mod_c=a%c;
-    for (int i=0;i<b;i++){
+    for (unsigned long long i=0;i<b;i++){
         ans = (ans*a_mod_c)%c;
     }
     cout<<ans;
